feat(Lotto) : 로또 model 클래스 구현
