+++ src/main/java/lotto/view/InputView.java
@@ -4,8 +4,6 @@ import camp.nextstep.edu.missionutils.Console;
 import lotto.exception.PriceToBuyLottoException;
 import lotto.model.PriceToBuyLotto;
 
-import static lotto.common.constant.ErrorMessage.*;
-
 public class InputView {
 
     public static PriceToBuyLotto getPriceToBuyLotto() {
